bump module version for 1.2.1 (v1213), [Gen 2] bootloaders, update CHANGELOG.md, update upgrade-downgrade & country-updown scripts

+++ user/tests/scripts/country-updown/country-updown.sh
@@ -53,8 +53,9 @@ export PHOTON_SP2_070="https://github.com/particle-iot/device-os/releases/downlo
 export PHOTON_SP1_101="https://github.com/particle-iot/device-os/releases/download/v1.0.1/system-part1-1.0.1-photon.bin"
 export PHOTON_SP2_101="https://github.com/particle-iot/device-os/releases/download/v1.0.1/system-part2-1.0.1-photon.bin"
 export PHOTON_SP1_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part1@1.1.0.bin"
-export PHOTON_SP2_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part1@1.1.0.bin"
-
+export PHOTON_SP2_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part2@1.1.0.bin"
+export PHOTON_SP1_121="https://github.com/particle-iot/device-os/releases/download/v1.2.1/photon-system-part1@1.2.1.bin"
+export PHOTON_SP2_121="https://github.com/particle-iot/device-os/releases/download/v1.2.1/photon-system-part2@1.2.1.bin"
 
 # instead of these set options,
 #   set script options
@@ -374,6 +375,13 @@ curl_all_required_system_parts() {
     curl --fail -L "$PHOTON_SP2_110" -o "system-part2-1.1.0-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
   fi
 
+  if [ ! -e "system-part1-1.2.1-photon.bin" ]; then
+    curl --fail -L "$PHOTON_SP1_121" -o "system-part1-1.2.1-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
+  fi
+  if [ ! -e "system-part2-1.2.1-photon.bin" ]; then
+    curl --fail -L "$PHOTON_SP2_121" -o "system-part2-1.2.1-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
+  fi
+
   if [ ! -e "tinker-v0.4.9-photon.bin" ]; then
     curl --fail -L "$PHOTON_TKR_049" -o "tinker-v0.4.9-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
   fi
@@ -604,14 +612,14 @@ fail() {
 #  87 |   JP2   |   050   | DEFAULT |   101      |   GB0   |   070/063/050   | wiped creds, re-enter
 #  88 |   JP2   |   050   | DEFAULT |   101      | DEFAULT |   070/063/050   | re-test, online
 # ----+---------+---------+---------+------------+---------+-----------------+----------------------
-#  89 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/050   | online
-#  90 |   JP2   |   050   | DEFAULT | 051/110    |   N/A   |   070/063/050   | online
-#  91 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/051   | online
-#  92 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/050   | online
-#  93 |   JP2   |   050   | DEFAULT | 051/110    |   JP2   |   070/063/050   | online
-#  94 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/051   | online
-#  95 |   JP2   |   050   | DEFAULT |   110      |   GB0   |   070/063/050   | wiped creds, re-enter
-#  96 |   JP2   |   050   | DEFAULT |   110      | DEFAULT |   070/063/050   | re-test, online
+#  89 |   JP2   |   050   | DEFAULT |   121      |   N/A   |   070/063/050   | online
+#  90 |   JP2   |   050   | DEFAULT | 051/121    |   N/A   |   070/063/050   | online
+#  91 |   JP2   |   050   | DEFAULT |   121      |   N/A   |   070/063/051   | online
+#  92 |   JP2   |   050   | DEFAULT |   121      |   JP2   |   070/063/050   | online
+#  93 |   JP2   |   050   | DEFAULT | 051/121    |   JP2   |   070/063/050   | online
+#  94 |   JP2   |   050   | DEFAULT |   121      |   JP2   |   070/063/051   | online
+#  95 |   JP2   |   050   | DEFAULT |   121      |   GB0   |   070/063/050   | wiped creds, re-enter
+#  96 |   JP2   |   050   | DEFAULT |   121      | DEFAULT |   070/063/050   | re-test, online
 # ----+---------+---------+---------+------------+---------+-----------------+----------------------
 # LAST TEST | ALL PASSED
 # -------------------------
@@ -3230,12 +3238,12 @@ fi
 
 if true; then
   # ------
-  # 110
+  # 121
   # ------
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 89 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/050   | online                 |"
+  echo "| 89 |   JP2   |   050   | DEFAULT |   121      |   N/A   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3247,11 +3255,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3273,7 +3283,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 90 |   JP2   |   050   | DEFAULT | 051/110    |   N/A   |   070/063/050   | online                 |"
+  echo "| 90 |   JP2   |   050   | DEFAULT | 051/121    |   N/A   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3289,11 +3299,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3315,7 +3327,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 91 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/051   | online                 |"
+  echo "| 91 |   JP2   |   050   | DEFAULT |   121      |   N/A   |   070/063/051   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3327,11 +3339,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3353,7 +3367,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 92 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/050   | online                 |"
+  echo "| 92 |   JP2   |   050   | DEFAULT |   121      |   JP2   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3365,11 +3379,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3395,7 +3411,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 93 |   JP2   |   050   | DEFAULT | 051/110    |   JP2   |   070/063/050   | online                 |"
+  echo "| 93 |   JP2   |   050   | DEFAULT | 051/121    |   JP2   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3411,11 +3427,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3441,7 +3459,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 94 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/051   | online                 |"
+  echo "| 94 |   JP2   |   050   | DEFAULT |   121      |   JP2   |   070/063/051   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3453,11 +3471,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3483,7 +3503,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 95 |   JP2   |   050   | DEFAULT |   110      |   GB0   |   070/063/050   | wiped creds, re-enter  |"
+  echo "| 95 |   JP2   |   050   | DEFAULT |   121      |   GB0   |   070/063/050   | wiped creds, re-enter  |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3495,11 +3515,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
@@ -3525,7 +3547,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 96 |   JP2   |   050   | DEFAULT |   110      | DEFAULT |   070/063/050   | re-test, online        |"
+  echo "| 96 |   JP2   |   050   | DEFAULT |   121      | DEFAULT |   070/063/050   | re-test, online        |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
 
   enter_ymodem
@@ -3541,11 +3563,13 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.2.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
-  sleep 30
+  # wait for 1.2.1 bootloader from SMH
+  sleep 10
+  try ymodem_boot 1.2.1 photon
+  sleep 10
 
   enter_dfu_mode
   try dfu_system 0.7.0 photon
